pyproject.toml
[tool.poetry]
name = "dpy-utils"
version = "0.1.0"
description = ""
authors = ["makubob <makupi@icloud.com>"]
license = "GPLv3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vvv"
testpaths = ["tests"]

[tool.poe.tasks]
test = { shell = "python -m pytest" }

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
"discord.py" = "^1.5.1"
aiounittest = "^1.4.0"
pytest = "^6.2.1"
poethepoet = "^0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
